48 harmonize packaging configuration and dependencies for 015

+++ dissmodel/models/sysdyn/sir.py
@@ -59,13 +59,13 @@ class SIR(Model):
     """
 
     #: Number of susceptible individuals.
-    susceptible: int
+    susceptible: float
 
     #: Number of infected individuals.
-    infected: int
+    infected: float
 
     #: Number of recovered individuals.
-    recovered: int
+    recovered: float
 
     #: Average number of steps an individual remains infectious.
     duration: int
